+++ src/main/kotlin/model/Day.kt
@@ -25,9 +25,9 @@ abstract class Day(
     override val from: LocalDate = LocalDate.now(),
     override val to: LocalDate = LocalDate.now(),
 
-    override val hours: Int,
+    override val hours: Double,
 
     @ElementCollection
-    override val days: List<Int>? = null
+    override val days: List<Double>? = null
 
 ) : Dayly, AbstractCodeEntity(id, code)
